cbr failed. 5th attemp with rar, last resort
The comic pages inside have been recompressed in photoshop and in win rar too. Probably the wrong order loading was due to a wrong order of decompression and the script was sorting by date of creation instead of name file order. But this is just a supposition, let's see the result.